requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.3
python-dotenv==1.0.1
h5py
numpy
scikit-learn
